codesprint-backend/requirements.txt
fastapi>=0.109.0
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.1
python-jose>=3.4.0
passlib==1.7.4
supabase==2.4.0
email-validator==2.1.0
